MOD: implements SPR(Single Responsibility Principle) in courseView package
